backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "walkthrough"
version = "0.1.0"
description = "SOP-to-Simulation Pipeline backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "google-cloud-storage>=2.18",
    "google-cloud-firestore>=2.19",
    "google-cloud-aiplatform>=1.74",
    "google-cloud-documentai>=2.32",
    "anthropic>=0.42",
    "python-multipart>=0.0.17",
    "sse-starlette>=2.1",
]

[dependency-groups]
dev = ["pyright>=1.1"]

[tool.hatch.build.targets.wheel]
packages = ["src/walkthrough"]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
